+++ source/AggregationResultReceiver/source/LocalMessageHub/Functions/PeekRequestListener.cs
@@ -16,6 +16,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Energinet.DataHub.Aggregations.LocalMessageHub.Storage;
 using Energinet.DataHub.MessageHub.Client.Peek;
 using Energinet.DataHub.MessageHub.Client.Storage;
 using Energinet.DataHub.MessageHub.Model.Extensions;
